Hotfix for inconsistent hashCode() and equals(Object) in Term hierarchy
Required by contract inherited from class Object.
Needed for use of Terms in hash-based data structures.

Current implementation provides only a hotfix (a constant value as the
shared hash code of all objects that inherit hashCode() from TermInherit
is not ideal); to be revisited for improved efficiency of hash-based
data structures on terms.